feat: better manual HTTPS server control
This is what was necessary to be able to run certbot externally to
refresh the TLS certificates and then restart just the HTTP server.
Restarting the whole process would require killing the JVM, which would
increase overall downtime.

A further thing I could do is have yet another future that runs once a
month, calls certbot itself, and then restarts the HTTP server if
certbot was successful.