fix: pre-sort `desc` searches by vote before calling `take`
In the particular case of just searching for `firefox`, ~400 results are
found. This is far too many, but because the `take` was previously being
called first, it was cutting off many popular packages, making them
absent from the results.

This also further reduces the max yielded results to 50 from 100. It's
taken as an assumption that in such large-yield cases, the user would
want to see the packages with the most votes anyway. Otherwise they will
have to choose wiser search terms.